# docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "JellyRick 🚀",
    "description": "API to post comments about Rick & Morty universe.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/comment/{id}": {
      "put": {
        "tags": [
          "comment"
        ],
        "summary": "Comment Put Route",
        "description": "route to update comment data",
        "operationId": "comment_put_route_api_v1_comment__id__put",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            },
            "name": "comment_id",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__comment__update_comment_router__CommentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "comment"
        ],
        "summary": "Comment Delete Route",
        "description": "route to delete comment data",
        "operationId": "comment_delete_route_api_v1_comment__id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            },
            "name": "comment_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/episode": {
      "get": {
        "tags": [
          "episode"
        ],
        "summary": "Episode List Route",
        "description": "route to get episode data from rick and morty universe",
        "operationId": "episode_list_route_api_v1_episode_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1,
              "type": "integer",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_str_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ]
      }
    },
    "/api/v1/episode/limit-offset": {
      "get": {
        "tags": [
          "episode"
        ],
        "summary": "Episode List Route",
        "description": "route to get episode data from rick and morty universe",
        "operationId": "episode_list_route_api_v1_episode_limit_offset_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Limit",
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "default": 50
            },
            "name": "limit",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Offset",
              "minimum": 0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_str_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "me"
            ]
          }
        ]
      }
    },
    "/api/v1/episode/{episode_id}": {
      "get": {
        "tags": [
          "episode"
        ],
        "summary": "Episode Unique Route",
        "description": "route to get episode data from rick and morty universe",
        "operationId": "episode_unique_route_api_v1_episode__episode_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "integer"
            },
            "name": "episode_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/comment/{comment_id}": {
      "get": {
        "tags": [
          "comment"
        ],
        "summary": "Comment Unique Route",
        "description": "route to get comment data from rick and morty universe",
        "operationId": "comment_unique_route_api_v1_comment__comment_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            },
            "name": "comment_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comment/episode/{episode_id}": {
      "post": {
        "tags": [
          "comment"
        ],
        "summary": "Comment Route",
        "description": "route to post comment data about Rick and Morty episodes.",
        "operationId": "comment_route_api_v1_comment_episode__episode_id__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Target Id",
              "type": "integer"
            },
            "name": "target_id",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__comment__post_comment_router__CommentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comment/character/{character_id}": {
      "post": {
        "tags": [
          "comment"
        ],
        "summary": "Comment Route",
        "description": "route to post comment data about Rick and Morty characters.",
        "operationId": "comment_route_api_v1_comment_character__character_id__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Target Id",
              "type": "integer"
            },
            "name": "target_id",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__comment__post_comment_router__CommentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comment/character-in-episode": {
      "post": {
        "tags": [
          "comment"
        ],
        "summary": "Comment Route With Parameters",
        "description": "route to post comment data about Rick and Morty character in episode.",
        "operationId": "comment_route_with_parameters_api_v1_comment_character_in_episode_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "integer"
            },
            "name": "episode_id",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "integer"
            },
            "name": "character_id",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__comment__post_comment_router__CommentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "LimitOffsetPage_str_": {
        "title": "LimitOffsetPage[str]",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0,
            "type": "integer"
          },
          "limit": {
            "title": "Limit",
            "minimum": 1,
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "minimum": 0,
            "type": "integer"
          }
        }
      },
      "Page_str_": {
        "title": "Page[str]",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0,
            "type": "integer"
          },
          "page": {
            "title": "Page",
            "minimum": 1,
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "minimum": 1,
            "type": "integer"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
      "app__routers__comment__post_comment_router__CommentBody": {
        "title": "CommentBody",
        "type": "object",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string",
            "default": "Nice comment example on episode !"
          }
        }
      },
      "app__routers__comment__update_comment_router__CommentBody": {
        "title": "CommentBody",
        "type": "object",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string",
            "default": "Nice comment update example !"
          }
        }
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {
              "me": "Read information about the current user.",
              "items": "Read items."
            },
            "tokenUrl": "/api/v1/login"
          }
        }
      }
    }
  }
}